# policies/rust/now-policy-api/openapi/now-policy-api.yaml
openapi: 3.1.0
info:
  title: Devolutions NOW Package Broker API
  description: HTTP API exposed by a Devolutions NOW package broker facade over a Windows named pipe.
  version: '1.0'
paths:
...
  /v1/policy:
    get:
      summary: Get active policy
      description: Returns the active parsed policy document. A 404 response means no active policy is configured.
      responses:
        default:
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: Response body for `GET /v1/policy`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
        '404':
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Replace the configured policy
      description: Reparses and revalidates the raw draft inside the write transaction, then atomically commits it only when the expected opaque store token and validation receipt still match.
      requestBody:
        description: Request body for `PUT /v1/policy`.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyReplacementRequest'
        required: true
      responses:
        '200':
          description: Response body for `PUT /v1/policy`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyReplacementResponse'
        '400':
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '501':
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/policy/management:
    get:
      summary: Get policy management state
      description: Atomically returns configured policy state and advisory write capability. Capability fields are UX guidance and are rechecked during replacement.
      responses:
        default:
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: Response body for `GET /v1/policy/management`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyManagementResponse'
  /v1/policy/validate:
    post:
      summary: Validate a policy draft
      description: Authoritatively validates raw draft JSON without discarding unknown fields. Validation findings are returned with HTTP 200; malformed envelopes use ErrorResponse.
      requestBody:
        description: Request body for `POST /v1/policy/validate`.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyValidationRequest'
        required: true
      responses:
        default:
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: Response body for `POST /v1/policy/validate`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyValidationResponse'
        '400':
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/package-operations/evaluate:
    post:
      summary: Evaluate package operation
      description: Evaluates a package operation against policy without requiring elevated execution.
      requestBody:
        description: Canonical request sent by a package broker client to the elevated broker.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageRequest'
        required: true
      responses:
        '200':
          description: Canonical response returned by the broker after evaluating a request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationResponse'
        '400':
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/package-operations/execute:
    post:
      summary: Execute package operation
      description: Evaluates a package operation and submits it to the implementation for execution when allowed.
      requestBody:
        description: Canonical request sent by a package broker client to the elevated broker.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageRequest'
        required: true
      responses:
        '200':
          description: Response returned after an execute request is evaluated and, when allowed, submitted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
        '400':
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/package-operations/get-status:
    post:
      summary: Query package operation status
      description: Returns the current status of a previously submitted package operation.
      requestBody:
        description: Request body for querying an operation status.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusRequest'
        required: true
      responses:
        '200':
          description: Response to a status query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/package-operations/cancel:
    post:
      summary: Cancel package operation
      description: 'Requests cancelation of a previously submitted package operation. Cancelation is asynchronous: poll the status endpoint until the operation reaches a terminal status.'
      requestBody:
        description: Request body for canceling a previously submitted operation.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelRequest'
        required: true
      responses:
        '200':
          description: |-
            Response to a cancel request.

             Cancelation is asynchronous and idempotent: the broker acknowledges the request by
             moving a non-terminal operation to `Canceling` and reports the resulting status.
             Clients should poll the status endpoint until the operation reaches a terminal
             status (`Canceled`, or `Completed`/`Failed` when the process ends first).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelResponse'
        '400':
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Generic error body returned for non-2xx responses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    ErrorResponse:
      description: Generic error body returned for non-2xx responses.
      type: object
      properties:
        Code:
          description: Machine-readable error code.
          allOf:
          - $ref: '#/components/schemas/ErrorCode'
        Details:
          description: Structured error details.
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
        Message:
          description: Human-readable summary.
          type: string
          maxLength: 2048
          minLength: 1
        ResponseKind:
          description: Response discriminator.
          allOf:
          - $ref: '#/components/schemas/ErrorResponseKind'
        ResponseVersion:
          description: Server-side API version used to construct the response.
          allOf:
          - $ref: '#/components/schemas/ApiVersion'
        Server:
          description: Server context.
          allOf:
          - $ref: '#/components/schemas/ServerContext'
        Validation:
          description: Current authoritative policy findings for management errors.
          anyOf:
          - $ref: '#/components/schemas/PolicyValidationResult'
          - enum:
            - null
            nullable: true
      required:
      - ResponseKind
      - ResponseVersion
      - Server
      - Code
      - Message
...
    PolicyManagementResponse:
      description: Response body for `GET /v1/policy/management`.
      type: object
      properties:
        Management:
          $ref: '#/components/schemas/PolicyManagementSnapshot'
        ResponseKind:
          $ref: '#/components/schemas/PolicyManagementResponseKind'
        ResponseVersion:
          $ref: '#/components/schemas/ApiVersion'
        Server:
          $ref: '#/components/schemas/ServerContext'
      additionalProperties: false
      required:
      - ResponseKind
      - ResponseVersion
      - Server
      - Management
...
    PolicyReplacementRequest:
      description: Request body for `PUT /v1/policy`.
      type: object
      properties:
        ConflictHandling:
          $ref: '#/components/schemas/PolicyConflictHandling'
        Draft:
          description: Raw draft JSON retained for transaction-time reparsing and revalidation.
        ExpectedStoreToken:
          $ref: '#/components/schemas/PolicyStoreToken'
        Operation:
          $ref: '#/components/schemas/PolicyReplacementOperation'
        RequestKind:
          $ref: '#/components/schemas/PolicyReplacementRequestKind'
        RequestVersion:
          $ref: '#/components/schemas/ApiVersion'
        ValidationReceipt:
          $ref: '#/components/schemas/PolicyValidationReceipt'
        WarningsAcknowledged:
          description: Explicit acknowledgement of every warning bound into the validation receipt.
          type: boolean
      additionalProperties: false
      required:
      - RequestKind
      - RequestVersion
      - ExpectedStoreToken
      - Operation
      - ConflictHandling
      - WarningsAcknowledged
      - Draft
      - ValidationReceipt
...
    PolicyReplacementResponse:
      description: Response body for `PUT /v1/policy`.
      type: object
      properties:
        Management:
          description: Newly observed management state and store token.
          allOf:
          - $ref: '#/components/schemas/PolicyManagementSnapshot'
        Policy:
          description: Exact committed active policy, including server-assigned metadata.
          allOf:
          - $ref: '#/components/schemas/PolicyDocument'
        ResponseKind:
          $ref: '#/components/schemas/PolicyReplacementResponseKind'
        ResponseVersion:
          $ref: '#/components/schemas/ApiVersion'
        Server:
          $ref: '#/components/schemas/ServerContext'
        Validation:
          description: Transaction-time validation result for the exact committed draft.
          allOf:
          - $ref: '#/components/schemas/PolicyValidationResult'
      additionalProperties: false
      required:
      - ResponseKind
      - ResponseVersion
      - Server
      - Policy
      - Validation
      - Management
...
    PolicyResponse:
      description: Response body for `GET /v1/policy`.
      type: object
      properties:
        Policy:
          description: Active parsed policy document.
          allOf:
          - $ref: '#/components/schemas/PolicyDocument'
        ResponseKind:
          description: Response discriminator.
          allOf:
          - $ref: '#/components/schemas/PolicyResponseKind'
        ResponseVersion:
          description: Server-side API version used to construct the response.
          allOf:
          - $ref: '#/components/schemas/ApiVersion'
        Server:
          description: Server context.
          allOf:
          - $ref: '#/components/schemas/ServerContext'
      required:
      - ResponseKind
      - ResponseVersion
      - Server
      - Policy
...
    PolicyValidationRequest:
      description: Request body for `POST /v1/policy/validate`.
      type: object
      properties:
        Draft:
          description: Raw draft JSON retained without dropping unknown members.
        RequestKind:
          $ref: '#/components/schemas/PolicyValidationRequestKind'
        RequestVersion:
          $ref: '#/components/schemas/ApiVersion'
      additionalProperties: false
      required:
      - RequestKind
      - RequestVersion
      - Draft
...
    PolicyValidationResponse:
      description: Response body for `POST /v1/policy/validate`.
      type: object
      properties:
        ResponseKind:
          $ref: '#/components/schemas/PolicyValidationResponseKind'
        ResponseVersion:
          $ref: '#/components/schemas/ApiVersion'
        Server:
          $ref: '#/components/schemas/ServerContext'
        Validation:
          $ref: '#/components/schemas/PolicyValidationResult'
      additionalProperties: false
      required:
      - ResponseKind
      - ResponseVersion
      - Server
      - Validation
...